server/python-server/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai==0.8.5